Отрефакторит filmPopupView
